Update canDelete logic for Product and Stock models
Refined the canDelete method in Product and Stock models to add additional checks: Product can only be deleted if it is not trashed and has more than one stock; Stock can only be deleted if it is not trashed and has at least one item. Also added items() relationship to Stock model.